refactor: separate validation logic to a middleware
 - middlewares/repoUrlValidator.js
