+++ utils.js
@@ -1,12 +1,15 @@
 const { cloneDeep } = require('lodash');
 const GIT = require('./constants/gitConstants');
 
-function hasGitExtension(repoName) {
-  return repoName.slice(-4) === GIT.GIT_EXTENSION;
-}
+const hasGitExtension = (repoName) => repoName.slice(-4) === GIT.GIT_EXTENSION;
 
-function changeBranchNameFormat(logList) {
-  return cloneDeep(logList).map((log) => {
+const getRepoName = (repoUrl) =>
+  hasGitExtension(repoUrl.split('/')[4])
+    ? repoUrl.split('/')[4].slice(0, -4)
+    : repoUrl.split('/')[4];
+
+const changeBranchNameFormat = (logList) =>
+  cloneDeep(logList).map((log) => {
     const splittedBranchName = log.branchName.split('/');
 
     return {
@@ -14,9 +17,9 @@ function changeBranchNameFormat(logList) {
       branchName: splittedBranchName[splittedBranchName.length - 1],
     };
   });
-}
 
 module.exports = {
   hasGitExtension,
+  getRepoName,
   changeBranchNameFormat,
 };
